Exported some enums and interfaces to character-data
