+++ src/api/player-details.ts
@@ -1,3 +1,5 @@
+import { Path, Element, PlayerData, PlayerDetails, Character, DisplayedCharacters } from "../utils/character-data";
+
 /**
  * This function returns the link to the player's data
  * @param uuid Player's uuid to fetch the data from
@@ -5,60 +7,6 @@
  */
 function player_api_link(uuid: string): string {
     return `https://api.mihomo.me/sr_info_parsed/${uuid}?lang=en`;
-}
-
-// Enum of acceptable paths
-enum Path {
-    "The Destruction",
-    "The Hunt",
-    "The Erudition",
-    "The Harmony",
-    "The Nihility",
-    "The Preservation",
-    "The Abundance"
-}
-
-// Enum of acceptable elements
-enum Element {
-    "Physical",
-    "Fire",
-    "Ice",
-    "Lightning",
-    "Wind",
-    "Quantum",
-    "Imaginary"
-}
-
-export interface PlayerData {
-    player: PlayerDetails,
-    characters: DisplayedCharacters
-}
-
-export interface PlayerDetails {
-    uuid: string,
-    nickname: string,
-    level: number,
-    world_level: number,
-    friend_count: number,
-    avatar_name: string,
-    signature: string,
-    light_cone_count: number,
-    avatar_count: number,
-    achievement_count: number
-}
-
-export interface DisplayedCharacters {
-    characters: Character[],
-}
-
-export interface Character {
-    name: string,
-    rarity: number,    // Rank (4* / 5*)
-    rank: number,      // Eidolon
-    level: number,
-    promotion: number, // Ascension
-    path: Path,
-    element: Element,
 }
 
 export default async function importPlayer(uuid: string) {
